refactor - moved common code to livekit/protocol

+++ bootstrap.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
 
-cd /tmp
-git clone https://github.com/magefile/mage
-cd mage
-go run bootstrap.go
-rm -rf /tmp/mage
+if ! command -v mage &> /dev/null
+then
+  pushd /tmp
+  git clone https://github.com/magefile/mage
+  cd mage
+  go run bootstrap.go
+  rm -rf /tmp/mage
+  popd
+fi
 
 if ! command -v mage &> /dev/null
 then
-  echo "Ensure ${GOPATH}/bin is in your \$PATH"
+  echo "Ensure `go env GOPATH`/bin is in your \$PATH"
+  exit 1
 fi
